+++ test/viking/model/class_properties/reflect_on_associations.js
@@ -2,7 +2,7 @@
     module("Viking.Model::reflectOnAssociations");
 
     test("::reflectOnAssociations() returns all the assocations", function() {
-        Ship = Viking.Model.extend({ hasMany: ['ships'], belongsTo: [['carrier', {model: 'Ship'}]] });
+        Ship = Viking.Model.extend({ hasMany: ['ships'], belongsTo: [['carrier', {modelName: 'Ship'}]] });
         ShipCollection = Viking.Collection.extend({ model: Ship });
         
         deepEqual(['carrier', 'ships'], _.map(Ship.reflectOnAssociations(), function(a) { return a.name; }));
@@ -12,7 +12,7 @@
     });
     
     test("::reflectOnAssociations(macro) returns all macro assocations", function() {
-        Ship = Viking.Model.extend({ hasMany: ['ships'], belongsTo: [['carrier', {model: 'Ship'}]] });
+        Ship = Viking.Model.extend({ hasMany: ['ships'], belongsTo: [['carrier', {modelName: 'Ship'}]] });
         ShipCollection = Viking.Collection.extend({ model: Ship });
         
         deepEqual(['ships'], _.map(Ship.reflectOnAssociations('hasMany'), function(a) { return a.name; }));